tests/wpt/tests/css/css-grid/grid-lanes/tentative/parsing/grid-lanes-shorthand-computed.html

<!DOCTYPE html>
<html>
<head>
  <title>CSS Grid Lanes: grid-lanes getComputedStyle()</title>
  <link rel="author" title="Yanling Wang" href="mailto:yanlingwang@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-grid-3">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
  <script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <script>
    test_computed_value("grid-lanes", '"a b" 1fr 2fr row');
    test_computed_value("grid-lanes", "8px column");
    test_computed_value("grid-lanes", '"a b c" 10% 20% 30% row-reverse', '"a b c" 10% 20% 30% row-reverse');
    test_computed_value("grid-lanes", '"a b" 10px 20px column-reverse');
    test_computed_value("grid-lanes", 'min-content row', 'min-content row');
    test_computed_value("grid-lanes", 'repeat(5, auto)', 'repeat(5, auto) normal');
  </script>
</body>
</html>
